# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NiChart API",
    "summary": "Backend API for the NiChart medical-imaging pipeline platform.",
    "description": "Provides endpoints for managing projects, uploading imaging data, running containerised processing pipelines (locally via Docker or on AWS Batch), and retrieving results. \n\n**Authentication**: In cloud mode every request requires a Cognito ID token in the ``Authorization: Bearer`` header. In local mode (``NICHART_EXECUTION_MODE=local``) authentication is bypassed.",
    "contact": {
      "name": "CBICA",
      "url": "https://github.com/CBICA/NiChart_Project"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/projects/{project_id}": {
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Delete a project",
        "description": "Permanently deletes the project and all its data. Irreversible.",
        "operationId": "delete_project_projects__project_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Project not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/files": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "List project files",
        "description": "Returns a flat listing of all visible files and directories in the project. Internal staging (``_upload/``) and working (``_working/``) directories are excluded.",
        "operationId": "list_files_projects__project_id__files_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryTree"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Access denied to this project.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Project or file not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or unsafe path.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Files"
        ],
        "summary": "Delete a file or directory",
        "description": "Permanently deletes the file or directory at ``path``.",
        "operationId": "delete_file_projects__project_id__files_delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Path relative to the project root.",
              "title": "Path"
            },
            "description": "Path relative to the project root."
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Access denied to this project.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Project or file not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or unsafe path.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/files/download": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "Download a file or directory",
        "description": "Download a single file, or a directory as a streaming zip archive when ``zip=true``. The ``path`` is relative to the project root and is validated to prevent traversal.",
        "operationId": "download_file_projects__project_id__files_download_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Path relative to the project root.",
              "title": "Path"
            },
            "description": "Path relative to the project root."
          },
          {
            "name": "zip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Stream target directory as a zip archive.",
              "default": false,
              "title": "Zip"
            },
            "description": "Stream target directory as a zip archive."
          }
        ],
        "responses": {
          "200": {
            "description": "File content or zip stream.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Access denied to this project.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Project or file not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or unsafe path.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/files/archive": {
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Bundle a multi-file/directory selection into one zip",
        "description": "Bundles an arbitrary set of files and/or directories into a single zip archive, preserving each entry's path relative to the project root. Useful for downloading a multi-selection from the file browser as one archive instead of one request per item. Named to avoid 'download' in the URL, which some ad-blocker/privacy extensions filter.",
        "operationId": "create_archive_projects__project_id__files_archive_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadZipRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Zip stream.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Access denied to this project.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Project or file not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or unsafe path.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/files/upload/nifti": {
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Upload NIfTI file(s) to staging",
        "description": "Accepts one or more ``.nii`` / ``.nii.gz`` files. Files land in the project staging area and the server returns its best-effort MRID and modality inference. Follow up with the commit endpoint to move them into the project.",
        "operationId": "upload_nifti_projects__project_id__files_upload_nifti_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_nifti_projects__project_id__files_upload_nifti_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NiftiStagingResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Access denied to this project.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Project or file not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or unsafe path.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_nifti_projects__project_id__files_upload_nifti_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_nifti_projects__project_id__files_upload_nifti_post"
      },
...
      "DirectoryTree": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/FileEntry"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "title": "DirectoryTree",
        "description": "Flat listing of all visible nodes in a project directory."
      },
...
      "DownloadZipRequest": {
        "properties": {
          "paths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Paths",
            "description": "Paths relative to the project root — a mix of files and directories is allowed. Each entry keeps its project-relative path inside the resulting archive, so the original directory structure is preserved."
          }
        },
        "type": "object",
        "required": [
          "paths"
        ],
        "title": "DownloadZipRequest",
        "description": "Request body for bundling multiple files/directories into a single zip."
      },
      "ErrorDetail": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable description of the error."
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "description": "Standard error response body."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}